# api/openapi-spec.yaml
openapi: 3.0.0
info:
  title: Images DB
  description: |
    This is the Images DB API based on the OpenAPI 3.0.0 specification.  
    Users are able to upload their images on the server and retrieve it.
  contact:
    name: Firas
    email: firasfj17@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:4000/v1
...
security:
  - jwt_auth: []
paths:
...
  /images:
    post:
      security:
        - jwt_auth: []
      tags:
        - images
      summary: Upload an image
      description: Upload an image providing its URL, title, description, and associated tags
      operationId: uploadImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  example: https://picsum.photos/id/407/1000/600
                title:
                  type: string
                  example: Delightful image
                description:
                  type: string
                  example: Taken in 2001
                tags:
                  type: array
                  items:
                    type: string
                  example: [red, green]
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully uploaded image!
                  imageId:
                    type: integer
                    example: 301
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"

    get:
      security:
        - jwt_auth: []
      tags:
        - images
      summary: Retrieve a batch of images
      description: Retrieve a batch of images with pagination and sorting
      operationId: getImages
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: Number of items to return
          example: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
          description: Number of items to skip
          example: 0
        - in: query
          name: sortBy
          schema:
            type: string
          description: Field to sort by
          example: id
        - in: query
          name: sortOrder
          schema:
            type: string
          description: Sort order (ASC or DESC)
          example: ASC
        - in: query
          name: showDeleted
          schema:
            type: string
          description: Whether to include deleted items (For admin)
          example: true
        - in: query
          name: showFlagged
          schema:
            type: string
          description: Whether to include flagged items (For admin)
          example: true
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          description: Comma-separated list of tags
          example: [nature]
        - in: query
          name: searchQuery
          schema:
            type: string
          description: Search query
          example: this
        - in: query
          name: searchColumn
          schema:
            type: string
          description: Column to search in
          example: title
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully fetched images!
                  fetched:
                    type: integer
                    example: 1
                  data:
                    $ref: "#/components/schemas/ImageArray"
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
                  totalImages:
                    type: integer
                    example: 1
                  totalNeededImages:
                    type: integer
                    example: 1
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: No images found
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      security:
        - jwt_auth: []
      tags:
        - images
      summary: Delete all your images
      description: Delete all of the images the signed-in user has uploaded
      operationId: deleteImages
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: All of your images deleted successfully.
                  deletedCount:
                    type: integer
                    example: 5
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: No images found to delete
        "500":
          $ref: "#/components/responses/ServerError"
  /images/{imageId}:
    get:
      security:
        - jwt_auth: []
      tags:
        - images
      summary: Retrieve image by ID
      operationId: getImageById
      parameters:
        - name: imageId
          in: path
          required: true
          description: ID of the image to retrieve
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageId:
                    type: integer
                    example: 123
                  imageUrl:
                    type: string
                    example: http://example.com/image.jpg
        "404":
          description: Image not found
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      security:
        - jwt_auth: []
      tags:
        - images
      summary: Delete image by ID
      description: Delete an image by its ID
      operationId: deleteImageById
      parameters:
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the image to delete
          example: 1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Image deleted
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Unauthorized to delete this image
        "404":
          description: Image not found
        "500":
          $ref: "#/components/responses/ServerError"

    patch:
      security:
        - jwt_auth: []
      tags:
        - images
      summary: Partially update image
      description: Partially update an image by its ID
      operationId: partiallyUpdateImage
      parameters:
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
          description: ID of the image to update
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                url:
                  type: string
                  example: https://picsum.photos/id/401/1000/600
                title:
                  type: string
                  example: A beautiful air balloon
                description:
                  type: string
                  nullable: true
                  example: This is quite a pic!
                ownerId:
                  type: integer
                  nullable: true
                  example: 1
                tags:
                  type: array
                  items:
                    type: string
                  nullable: true
                  example: [red, blue]
                isFlagged:
                  type: boolean
                  nullable: true
                  example: false
                createdAt:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  example: 2023-05-04
                updatedAt:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  example: 2023-05-04
                destroyTime:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  nullable: true
                  example: null

      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Image details updated successfully.
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Unauthorized to update this image
        "404":
          description: Image not found
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      security:
        - jwt_auth: []
      tags:
        - images
      summary: Update an image
      description: Completely update an image by its ID
      operationId: updateImage
      parameters:
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
          description: ID of the image to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                url:
                  type: string
                  example: https://picsum.photos/id/401/1000/600
                title:
                  type: string
                  example: A beautiful air balloon
                description:
                  type: string
                  nullable: true
                  example: This is quite a pic!
                ownerId:
                  type: integer
                  nullable: true
                  example: 1
                tags:
                  type: array
                  items:
                    type: string
                  nullable: true
                  example: [red, blue]
                isFlagged:
                  type: boolean
                  nullable: true
                  example: false
                createdAt:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  example: 2023-05-04
                updatedAt:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  example: 2023-05-04
                destroyTime:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  nullable: true
                  example: null
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Image details updated successfully
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Only an admin can send the PUT request
        "404":
          description: Image not found
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  schemas:
...
    Image:
      type: object
      properties:
        id:
          type: integer
          example: 1
        url:
          type: string
          example: https://picsum.photos/id/401/1000/600
        title:
          type: string
          maxLength: 65
          example: A beautiful air balloon
        description:
          type: string
          example: This is quite a pic!
        ownerId:
          type: integer
          example: 1
        tags:
          type: array
          items:
            type: string
          example: [red, blue]
        isFlagged:
          type: boolean
          example: false
        createdAt:
          type: string
          oneOf:
            - format: date
            - format: date-time
          example: 2023-05-04
        updatedAt:
          type: string
          oneOf:
            - format: date
            - format: date-time
          example: 2023-05-04
        destroyTime:
          type: string
          oneOf:
            - format: date
            - format: date-time
          nullable: true
          example: null
      required:
        - url
        - title
...
    ImageArray:
      description: Array of Images
      type: array
      items:
        $ref: "#/components/schemas/Image"
    HATEOASLink:
      type: object
      properties:
        rel:
          type: string
        method:
          type: string
        href:
          type: string
        description:
          type: boolean
    HATEOASLinkArray:
      description: Array of HATEOAS links
      type: array
      items:
        $ref: "#/components/schemas/HATEOASLink"
  responses:
    UnauthorizedError:
      description: Token not provided or invalid
    InvalidInputError:
      description: Invalid input
    ServerError:
      description: Internal server error occurred.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
  securitySchemes:
    jwt_auth:
      type: apiKey
      in: cookie
      name: jwt